enostr: rename Event to Note
we will likely replace Note with nostrdb::Note in the future,
this just helps with that transition
